Android Base Project
This includes the correct config to build an Android build but not tested yet on Linux.

Req:
Flutter SDK
Dart Plugin

This won't work unless you set up the variables for Flutter.

This is just a base UI with the basic configuration to start bringing over what we've completed on

https://github.com/NiRuX/Snippy-deprecated